AMLII-1948 - restructure existing tests in to new file structure
udp_test.go components that will require a udp connection moved to
udp_intergration_test.go

server_test.go utility functions moved to server_util_test.go

server_test.go and server_uds_test.go tests that require socket
interactions moved to server_integration_test.go

server-test.go functionality that test message parsing rather than
configuration/initialization moved to server_worker_test.go

remaining test in server_uds_test.go consolidated into server_test.go
